Register form and it's validation;

+++ src/main/java/in/nit/rohit/repo/SpecializationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import in.nit.rohit.entity.Specialization;
 
-public interface SpecializationRepository extends JpaRepository<Specialization,Integer> {
+public interface SpecializationRepository extends JpaRepository<Specialization,Long> {
 
 }
